% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnsw_download_map.R
\name{bnsw_download_map}
\alias{bnsw_download_map}
\title{Download a map from a given URL}
\usage{
bnsw_download_map(url, dest_path)
}
\arguments{
\item{url}{URL to given download}

\item{dest_path}{where to save the download}
}
\value{
character vector, path to downloaded file
}
\description{
Download a map from a given URL
}
