% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnsw_get_map.R
\name{bnsw_get_map}
\alias{bnsw_get_map}
\title{Import NSW map into R session}
\usage{
bnsw_get_map(
  boundary = c("lga2016", "lga2019", "lhd2010", "poa2016", "gccsa2016"),
  geojson_extract_location,
  return_sf_object = TRUE
)
}
\arguments{
\item{boundary}{character, code for selected boundary}

\item{geojson_extract_location}{character, a path for extracted geojson. If missing a \code{tempdir()} will be used}

\item{return_sf_object}{boolean, whether to return the map as an \code{sf} object}
}
\description{
This function returns a chosen map an \code{sf} object. The maps are stored as compressed
geojsons within this package. You can chose whether to extract that geojson to a
chosen destination (geojson_extract_location), or allow the function to extract to a temporary
folder for import of the geojson into R using \code{\link[sf:st_read]{sf::st_read()}}
}
\examples{
{
# Import LGA19 as sf object to R session
lga2019 <- bnsw_get_map('lga2019')

# Extract LGA19 geojson to working directory
lga2019 <- bnsw_get_map('lga2019', return_sf_object = FALSE)
}
}
