% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnsw_unzip.R
\name{bnsw_unzip}
\alias{bnsw_unzip}
\title{Unzip/extract a downloaded boundary archive}
\usage{
bnsw_unzip(zip, overwrite = FALSE)
}
\arguments{
\item{zip}{path to zip file}

\item{overwrite}{boolean, whether to overwrite files if already present}
}
\value{
character vector, path to extracted files
}
\description{
Unzip/extract a downloaded boundary archive
}
