% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bnsw_lga2019_code_names}
\alias{bnsw_lga2019_code_names}
\alias{bnsw_lga2016_code_names}
\title{Boundary metadata}
\source{
\url{https://www.todo.com.au}
}
\usage{
bnsw_lga2019_code_names

bnsw_lga2016_code_names
}
\description{
The columns associated with boundaries coordinates, such as codes and names
}
\details{
A data frame with some rows and columns
\describe{
\item{code, name}{Codes, names and other relevant things.}
}
}
\section{Functions}{
\itemize{
\item \code{bnsw_lga2019_code_names}: LGA 2019 codes and names

\item \code{bnsw_lga2016_code_names}: LGA 2016 codes and names

}}
\keyword{datasets}
