% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_maps.R
\name{list_maps}
\alias{list_maps}
\title{List map files associated with boundariesNSW}
\usage{
list_maps(full.names = FALSE)
}
\arguments{
\item{full.names}{a logical value. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.}
}
\description{
A simple wrapper to \code{list.files()} of the packages external data
}
