% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_map.R
\name{simplify_map}
\alias{simplify_map}
\title{Simplify a geometery}
\usage{
simplify_map(map, keep_proportion = 0.1)
}
\arguments{
\item{map}{sf object, sf data.frame with column geometry}

\item{keep_proportion}{number, proportion of map to keep. Must be between 0 - 1}
}
\description{
Sometimes a default shape file is of too high resolution for simple plotting.
This function uses \code{rmapshaper} to simplify map geometry.
}
